Components: Extract a reusable PostTrash component
