+++ editor/post-trash/style.scss
@@ -1,0 +1,10 @@
+.editor-post-trash {
+	margin-left: auto !important;
+	text-decoration: underline;
+
+	.dashicon {
+		display: inline-block;
+		vertical-align: middle;
+		margin: -3px -4px 0 10px;
+	}
+}
